+++ mmocr/core/deployment/__init__.py
@@ -1,5 +1,5 @@
-from .wrappers import (ONNXRuntimeDetector, ONNXRuntimeRecognizer,
-                       TensorRTDetector, TensorRTRecognizer)
+from .deploy_utils import (ONNXRuntimeDetector, ONNXRuntimeRecognizer,
+                           TensorRTDetector, TensorRTRecognizer)
 
 __all__ = [
     'ONNXRuntimeRecognizer', 'ONNXRuntimeDetector', 'TensorRTDetector',
